improve error handling
