+++ openml/_api_calls.py
@@ -1,6 +1,7 @@
 # License: BSD 3-Clause
 
 import time
+import hashlib
 import logging
 import requests
 import xmltodict
@@ -8,7 +9,7 @@ from typing import Dict
 
 from . import config
 from .exceptions import (OpenMLServerError, OpenMLServerException,
-                         OpenMLServerNoResult)
+                         OpenMLServerNoResult, OpenMLHashException)
 
 
 def _perform_api_call(call, request_method, data=None, file_elements=None):
@@ -54,11 +55,7 @@ def _perform_api_call(call, request_method, data=None, file_elements=None):
     else:
         response = _read_url(url, request_method, data)
 
-    if response.status_code != 200:
-        raise _parse_server_exception(response, url, file_elements=file_elements)
-    elif 'Content-Encoding' not in response.headers or \
-            response.headers['Content-Encoding'] != 'gzip':
-        logging.warning('Received uncompressed content from OpenML for {}.'.format(url))
+    _check_response(response, url, file_elements)
 
     logging.info(
         '%.7fs taken for [%s] request for the URL %s',
@@ -67,6 +64,75 @@ def _perform_api_call(call, request_method, data=None, file_elements=None):
         url,
     )
     return response.text
+
+
+def _download_text_file(source: str,
+                        output_path: str,
+                        md5_checksum: str = None,
+                        exists_ok: bool = True,
+                        encoding: str = 'utf8',
+                        ) -> None:
+    """ Download the text file at `source` and store it in `output_path`.
+
+    By default, do nothing if a file already exists in `output_path`.
+    The downloaded file can be checked against an expected md5 checksum.
+
+    Parameters
+    ----------
+    source : str
+        url of the file to be downloaded
+    output_path : str
+        full path, including filename, of where the file should be stored.
+    md5_checksum : str, optional (default=None)
+        If not None, should be a string of hexidecimal digits of the expected digest value.
+    exists_ok : bool, optional (default=True)
+        If False, raise an FileExistsError if there already exists a file at `output_path`.
+    encoding : str, optional (default='utf8')
+        The encoding with which the file should be stored.
+    """
+    try:
+        with open(output_path, encoding=encoding):
+            if exists_ok:
+                return
+            else:
+                raise FileExistsError
+    except FileNotFoundError:
+        pass
+
+    logging.info('Starting [%s] request for the URL %s', 'get', source)
+    start = time.time()
+    response = _read_url(source, request_method='get')
+    _check_response(response, source, None)
+    downloaded_file = response.text
+
+    if md5_checksum is not None:
+        md5 = hashlib.md5()
+        md5.update(downloaded_file.encode('utf-8'))
+        md5_checksum_download = md5.hexdigest()
+        if md5_checksum != md5_checksum_download:
+            raise OpenMLHashException(
+                'Checksum {} of downloaded file is unequal to the expected checksum {}.'
+                .format(md5_checksum_download, md5_checksum))
+
+    with open(output_path, "w", encoding=encoding) as fh:
+        fh.write(downloaded_file)
+
+    logging.info(
+        '%.7fs taken for [%s] request for the URL %s',
+        time.time() - start,
+        'get',
+        source,
+    )
+
+    del downloaded_file
+
+
+def _check_response(response, url, file_elements):
+    if response.status_code != 200:
+        raise _parse_server_exception(response, url, file_elements=file_elements)
+    elif 'Content-Encoding' not in response.headers or \
+            response.headers['Content-Encoding'] != 'gzip':
+        logging.warning('Received uncompressed content from OpenML for {}.'.format(url))
 
 
 def _file_id_to_url(file_id, filename=None):
